Week-1/3_basic_rag_system/requirements.txt
chromadb
pypdf2
python-docx
sentence-transformers
streamlit